+++ packages/@tailwindcss-upgrade/src/template/codemods/automatic-var-injection.test.ts
@@ -1,12 +1,11 @@
 import { __unstable__loadDesignSystem } from '@tailwindcss/node'
 import { expect, test } from 'vitest'
-import { printCandidate } from '../candidates'
 import { automaticVarInjection } from './automatic-var-injection'
 
 test.each([
   // Arbitrary candidates
   ['[color:--my-color]', '[color:var(--my-color)]'],
-  ['[--my-color:red]', null],
+  ['[--my-color:red]', '[--my-color:red]'],
   ['[--my-color:--my-other-color]', '[--my-color:var(--my-other-color)]'],
 
   // Arbitrary values for functional candidates
@@ -38,23 +37,23 @@ test.each([
   ['supports-[_--test]:flex', 'supports-[--test]:flex'],
 
   // Some properties never had var() injection in v3.
-  ['[scroll-timeline-name:--myTimeline]', null],
-  ['[timeline-scope:--myScope]', null],
-  ['[view-timeline-name:--myTimeline]', null],
-  ['[font-palette:--myPalette]', null],
-  ['[anchor-name:--myAnchor]', null],
-  ['[anchor-scope:--myScope]', null],
-  ['[position-anchor:--myAnchor]', null],
-  ['[position-try-options:--myAnchor]', null],
-  ['[scroll-timeline:--myTimeline]', null],
-  ['[animation-timeline:--myAnimation]', null],
-  ['[view-timeline:--myTimeline]', null],
-  ['[position-try:--myAnchor]', null],
+  ['[scroll-timeline-name:--myTimeline]', '[scroll-timeline-name:--myTimeline]'],
+  ['[timeline-scope:--myScope]', '[timeline-scope:--myScope]'],
+  ['[view-timeline-name:--myTimeline]', '[view-timeline-name:--myTimeline]'],
+  ['[font-palette:--myPalette]', '[font-palette:--myPalette]'],
+  ['[anchor-name:--myAnchor]', '[anchor-name:--myAnchor]'],
+  ['[anchor-scope:--myScope]', '[anchor-scope:--myScope]'],
+  ['[position-anchor:--myAnchor]', '[position-anchor:--myAnchor]'],
+  ['[position-try-options:--myAnchor]', '[position-try-options:--myAnchor]'],
+  ['[scroll-timeline:--myTimeline]', '[scroll-timeline:--myTimeline]'],
+  ['[animation-timeline:--myAnimation]', '[animation-timeline:--myAnimation]'],
+  ['[view-timeline:--myTimeline]', '[view-timeline:--myTimeline]'],
+  ['[position-try:--myAnchor]', '[position-try:--myAnchor]'],
 ])('%s => %s', async (candidate, result) => {
   let designSystem = await __unstable__loadDesignSystem('@import "tailwindcss";', {
     base: __dirname,
   })
 
-  let migrated = automaticVarInjection(designSystem, designSystem.parseCandidate(candidate)[0]!)
-  expect(migrated ? printCandidate(migrated) : migrated).toEqual(result)
+  let migrated = automaticVarInjection(designSystem, candidate)
+  expect(migrated).toEqual(result)
 })