+++ packages/@tailwindcss-upgrade/src/template/codemods/bg-gradient.ts
@@ -1,0 +1,20 @@
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+import { printCandidate } from '../candidates'
+
+const DIRECTIONS = ['t', 'tr', 'r', 'br', 'b', 'bl', 'l', 'tl']
+
+export function bgGradient(designSystem: DesignSystem, rawCandidate: string): string {
+  for (let candidate of designSystem.parseCandidate(rawCandidate)) {
+    if (candidate.kind === 'static' && candidate.root.startsWith('bg-gradient-to-')) {
+      let direction = candidate.root.slice(15)
+
+      if (!DIRECTIONS.includes(direction)) {
+        continue
+      }
+
+      candidate.root = `bg-linear-to-${direction}`
+      return printCandidate(candidate)
+    }
+  }
+  return rawCandidate
+}